[BY-211] Feat: Type buttons selectable
